Turn legal-to-review into more generic ?-to-review app
'legal-to-review' is just an example of broader concept of
commenting/requesting review from team(s) when certain file (or
group of files) get modified. It is not about developing another
actions but rather providing simple alternative when actions
are just to heavy for simple, modified-filename-based flow.

Changes:
1. both 'legalFileRegExp' and 'legalTeam' parameters of
'.github/config.yml' were dropped

2. list of review criteria objects is read from '.github/config.yml'
and it is defined as:

  reviewCriteria:
    - configName:
      name: 'display-name'
      regexp: 'pattern-to-search-for'
      teams?:
        - 'optional-reviewers-team-name'
    ...

3. by default if files doesn't exist or it doesn't contain
'reviewCriteria' definition 'legal-to-review' configuration is
assumed:

  - reviewCriteria:
    - legal:
      name: 'legal-to-review'
      regexp: '(licen(s|c)e)|(copyright)|(code.?of.?conduct)'